MISUV-5416v2 - Refactor controller name
