+++ conf/app.routes
@@ -167,10 +167,10 @@ POST         /agents/income-sources/cease/uk-property-declare             contro
 GET         /income-sources/cease/uk-property-end-date                    controllers.incomeSources.cease.DateUKPropertyCeasedController.show(origin: Option[String] ?= None)
 GET         /agents/income-sources/cease/uk-property-end-date             controllers.incomeSources.cease.DateUKPropertyCeasedController.showAgent()
 
-GET          /income-sources/add/business-name                            controllers.AddBusinessController.show()
-GET          /agents/income-sources/add/business-name                     controllers.AddBusinessController.showAgent()
-POST         /income-sources/add/business-name                            controllers.AddBusinessController.submit()
-POST         /agents/income-sources/add/business-name                     controllers.AddBusinessController.submitAgent()
+GET          /income-sources/add/business-name                            controllers.AddBusinessNameController.show()
+GET          /agents/income-sources/add/business-name                     controllers.AddBusinessNameController.showAgent()
+POST         /income-sources/add/business-name                            controllers.AddBusinessNameController.submit()
+POST         /agents/income-sources/add/business-name                     controllers.AddBusinessNameController.submitAgent()
 
 # Income Sources - TO BE IMPLEMENTED
 GET          /income-sources/add/uk-property-business-start-date                                        controllers.AddUkPropertyBusinessStartDate.show()
